+++ src/main/java/communication/messages/LobbyFullNotification.java
@@ -1,0 +1,10 @@
+
+package communication.messages;
+
+public class LobbyFullNotification {
+    private String messageType = "LobbyFullNotification";
+
+    public String getMessageType() {
+        return messageType;
+    }
+}