+++ src/main/java/view/BackgroundPanel.java
@@ -1,0 +1,19 @@
+package view;
+
+import javax.swing.*;
+import java.awt.*;
+
+public class BackgroundPanel extends JPanel {
+    private Image backgroundImage;
+
+    public BackgroundPanel(Image image) {
+        this.backgroundImage = image;
+        setLayout(new BorderLayout());
+    }
+
+    @Override
+    protected void paintComponent(Graphics g) {
+        super.paintComponent(g);
+        g.drawImage(backgroundImage, 0, 0, getWidth(), getHeight(), this);
+    }
+}