+++ CHANGELOG.md
@@ -11,25 +11,38 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Race code
 - GameScreen Class in View Package
 - Typing Player (instead of Typer)
+- "integration" branch
+- ClientWindow
+- LobbyFullNotification
+- PlayerListUpdateNotification
+- Methods in _ClientController_: HandlePlayerListUpdate(), HandleLobbyFull()
+- Methods in _GameServer_: BroadcastPlayerListUpdate(), BroadcastMessages(), BroadcastLobbyFull()
+- New Text samples in _Text Class_
 
 ### Changed
-GUI Class(with Main Method)
+- GUI Class(with Main Method)
+- GameStartNotification(added getters and setters)
+- PlayerJoinedNotification(getters and setters)
+
+
 ### Removed
-- LoginWindow Class
-- ResultWindow Class
-- GameWindow Class
-- Typer Class
+- LoginWindow Class(instead moved them to view package)
+- ResultWindow Class(instead moved them to view package)
+- GameWindow Class(instead moved them to view package)
+- Typer Class(instead made TypingPlayer)
 - 
 ### Fixed
 
 - Game Interface
 - Player Interface
+- GameState
 
 
 
 # NETWORK
 ## 22.Jun.2024
 ### Added
+
 - new class: messages, request for server and notis for client
 
 ## 24.Jun.2024
@@ -39,6 +52,9 @@ GUI Class(with Main Method)
 - GameServer
 - ConnectionManager
 - synchronized the processMessage class
+
+- new class: message, 
+- new class : PlayerListUpdateNotification in Messages package
 
 ### Changed
 - Changed the Interface of server and client into two separated interfaces
@@ -50,4 +66,11 @@ GUI Class(with Main Method)
 - Changed the Message class into  MessageType class, no more need to extends from this class
 
 ### Removed
-- all extends are removed 
+
+- all extends are removed
+
+### Fixed
+- GameServer
+- GameClient
+- ClientController
+